:recycle: Break travel desk hotels edit table out of edit card.
Why? Easier to reuse and customize when app components map to vuetify components.
