+++ web/src/components/travel-desk-hotels/TravelDeskHotelsEditCard.vue
@@ -15,40 +15,14 @@
       />
     </v-card-title>
     <v-card-text>
-      <TravelDeskHotelsDataTable
-        ref="travelDeskHotelsDataTable"
+      <TravelDeskHotelsEditDataTable
+        ref="travelDeskHotelsEditDataTable"
         :where="whereClause"
-      >
-        <template #top>
-          <TravelDeskHotelEditDialog
-            ref="editDialog"
-            :min-date="minDate"
-            :max-date="maxDate"
-            :flight-start="earliestFlightDate"
-            :flight-end="latestFlightDate"
-            @saved="refreshTravelDeskHotels"
-          />
-        </template>
-        <template #item.actions="{ item }">
-          <div class="d-flex justify-end">
-            <v-btn
-              title="Edit"
-              icon
-              color="blue"
-              @click="showEditDialog(item)"
-              ><v-icon>mdi-pencil</v-icon></v-btn
-            >
-            <v-btn
-              :loading="isLoading"
-              title="Delete"
-              icon
-              color="red"
-              @click="deleteHotel(item)"
-              ><v-icon>mdi-close</v-icon></v-btn
-            >
-          </div>
-        </template>
-      </TravelDeskHotelsDataTable>
+        :min-date="minDate"
+        :max-date="maxDate"
+        :flight-start="earliestFlightDate"
+        :flight-end="latestFlightDate"
+      />
     </v-card-text>
   </v-card>
 </template>
@@ -56,14 +30,11 @@
 <script setup>
 import { computed, ref, toRefs } from "vue"
 
-import blockedToTrueConfirm from "@/utils/blocked-to-true-confirm"
-import travelDeskHotelsApi from "@/api/travel-desk-hotels-api"
 import useTravelAuthorization from "@/use/use-travel-authorization"
 import useTravelDeskFlightRequests from "@/use/use-travel-desk-flight-requests"
 
 import TravelDeskHotelCreateDialog from "@/components/travel-desk-hotels/TravelDeskHotelCreateDialog.vue"
-import TravelDeskHotelEditDialog from "@/components/travel-desk-hotels/TravelDeskHotelEditDialog.vue"
-import TravelDeskHotelsDataTable from "@/components/travel-desk-hotels/TravelDeskHotelsDataTable.vue"
+import TravelDeskHotelsEditDataTable from "@/components/travel-desk-hotels/TravelDeskHotelsEditDataTable.vue"
 
 const props = defineProps({
   travelDeskTravelRequestId: {
@@ -99,33 +70,11 @@ const {
   refresh: refreshFlightRequests,
 } = useTravelDeskFlightRequests(travelDeskFlightRequestsQuery)
 
-/** @type {import("vue").Ref<InstanceType<typeof TravelDeskHotelEditDialog> | null>} */
-const editDialog = ref(null)
-
-function showEditDialog(hotel) {
-  editDialog.value?.show(hotel)
-}
-
-const isLoading = ref(false)
 /** @type {import("vue").Ref<InstanceType<typeof TravelDeskHotelsDataTable> | null>} */
-const travelDeskHotelsDataTable = ref(null)
+const travelDeskHotelsEditDataTable = ref(null)
 
 async function refreshTravelDeskHotels() {
-  return travelDeskHotelsDataTable.value?.refresh()
-}
-
-async function deleteHotel(hotel) {
-  if (!blockedToTrueConfirm("Are you sure you want to remove this hotel?")) return
-
-  isLoading.value = true
-  try {
-    await travelDeskHotelsApi.delete(hotel.id)
-    await refreshTravelDeskHotels()
-  } catch (error) {
-    console.error(error)
-  } finally {
-    isLoading.value = false
-  }
+  return travelDeskHotelsEditDataTable.value?.refresh()
 }
 
 defineExpose({
